remove non-linked min api and replace with submodules

+++ source/min-api/include/c74_min_operator_mc.h
@@ -57,7 +57,7 @@ namespace c74::min {
         /// It is called internally any time the dsp chain containing your object is compiled.
         /// @param	a_vector_size	A new vector size with which your object will be updated.
 
-        void vector_size(const double a_vector_size) {
+        void vector_size(const int a_vector_size) {
             m_vector_size = a_vector_size;
         }
 
@@ -65,7 +65,7 @@ namespace c74::min {
         /// Return the current vector size for this object's signal chain.
         /// @return	The vector size in samples.
 
-        double vector_size() const {
+        int vector_size() const {
             return m_vector_size;
         }
